Reorganized logic folder
