+++ src/core/logic/Strategies/moving_average.py
