+++ src/core/logic/Strategies/strategy.py
